+++ front-end/src/components/Userlist.jsx
@@ -1,6 +1,7 @@
 import React, {useState, useEffect} from 'react'
 import axios from 'axios';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 const Userlist = () => {
     const [users, setUsers] = useState([]);
@@ -17,35 +18,63 @@ const Userlist = () => {
     };
 
     const deleteUser = async (userId) => {
-        await axios.delete(`http://localhost:5000/users/${userId}`);
-        getUsers();
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#d33',
+        cancelButtonColor: '#3085d6',
+        confirmButtonText: 'Yes, delete it!',
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          await axios.delete(`http://localhost:5000/users/${userId}`);
+          Swal.fire({
+            icon: 'success',
+            title: 'User Deleted',
+            showConfirmButton: false,
+            timer: 1500,
+          });
+          getUsers();
+        }
+      });
     };
   return (
-    <div>
-        <h1 className='title'>Users</h1>
-     <h2 className='subtitle'> List of Users</h2>
+    <div className='bg-[#1e252b] w-screen h-screen'>
+        <h1 className='text-white text-3xl font-semibold'>Users</h1>
+     <h2 className='text-white text-xl font-semibold mb-4'> List of Users</h2>
     <Link to="/users/add" className='button is-primary mb-2'>Add new</Link>
-     <table className='table is-striped is-fullwidth'>
+     <table className='table is-striped is-fullwidth' style={{ backgroundColor: "#1e252b", color: "white" }}>
         <thead>
             <tr>
-                <th>No</th>
-                <th>Name</th>
-                <th>Email</th>
-                <th>Role</th>
-                <th>Actions</th>
+                <th>
+                  <p className='text-white'>No</p>
+                </th>
+                <th>
+                  <p className='text-white'>Name</p>
+                </th>
+                <th>
+                  <p className='text-white'>Email</p>
+                </th>
+                <th>
+                  <p className='text-white'>Role</p>
+                </th>
+                <th>
+                  <p className='text-white'>Actions</p>
+                </th>
             </tr>
         </thead>
         <tbody>
             {users.map((user, index)=>(
-              <tr key={user.uuid}>
-                <td>{index + 1}</td>
-                <td>{user.name}</td>
-                <td>{user.email}</td>
-                <td>{user.role}</td>
-                <td>
+              <tr className='bg-[#1e252b] text-white' key={user.uuid}>
+                <td className='bg-[#1e252b] text-white'>{index + 1}</td>
+                <td className='bg-[#1e252b] text-white'>{user.name}</td>
+                <td className='bg-[#1e252b] text-white'>{user.email}</td>
+                <td className='bg-[#1e252b] text-white'>{user.role}</td>
+                <td className='bg-[#1e252b] text-white'>
                 <Link
                   to={`/users/edit/${user.uuid}`}
-                  className="button is-small is-info"
+                  className="button is-small is-info mb-2 mr-2"
                 >
                   Edit
                 </Link>
